Move useLayerConfig to hooks
Move useLayerConfig to hooks
