+++ src/View/Board/RenderNG/Page/Content/index.tsx
@@ -1,9 +1,9 @@
 import React, { memo } from "react"
 import { useGState } from "state"
+import { useLayerConfig } from "hooks"
 import { Canvas } from "../index.styled"
 import { useRender } from "./useRender"
 import { PageProps } from "../index.types"
-import { useLayerConfig } from "../useLayerConfig"
 
 export const Content: React.FC<PageProps> = memo(({ page, pageOffset }) => {
     const canvasRef = React.useRef<HTMLCanvasElement>(null)
