+++ src/hooks/index.ts
@@ -3,3 +3,4 @@ export * from "./useInitialDialog"
 export * from "./useWindowResize"
 export * from "./useLiveStroke"
 export * from "./useViewControl"
+export * from "./useLayerConfig"
